Refactor and move crud.py to crud folder
